spilt navbar links into 2 components
